Move aabb cache into a separte file
